+++ app/src/main/java/fafour/projectthaiboxing/Total.java
@@ -5,8 +5,8 @@ package fafour.projectthaiboxing;
  */
 
 public class Total {
-    public static  int totalBuyItem(){
-        int total = 0;
+    public static  double totalBuyItem(){
+        double total = 0.0;
         for(int count=0; count < MainActivity.listBuy.size(); count++){
             total = total + (MainActivity.listBuy.get(count).accessoriesNum *
                     MainActivity.listBuy.get(count).accessoriesSale );
